django-auth/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
idna==2.10
Markdown==3.4.1
mysqlclient==1.4.6
pycodestyle==2.9.1
PyJWT==1.7.1
pyotp==2.6.0
python-decouple==3.6
pytz==2022.6
requests==2.27.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
